pyproject.toml
[tool.poetry]
name = "essembeh-tools"
version = "0.1.0"
description = "Collection of command line tools written in Python"
homepage = "https://github.com/essembeh/essembeh-tools"
authors = ["Sébastien MB <seb@essembeh.org>"]
license = "Apache-2.0"
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Utilities"
]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.9"
colorama = "^0.4.5"
python-magic = "^0.4.27"


[tool.poetry.dev-dependencies]
black = "^22.8.0"
pylint = "^2.15.3"
pytest = "^7.1.3"
pytest-dotenv = "^0.5.2"
pytest-cov = "^3.0.0"


[tool.poetry.scripts]
helloworld = "essembeh_tools.cli.helloworld:run"
dispatch = "essembeh_tools.cli.dispatch:run"
hrenamer = "essembeh_tools.cli.hrenamer:run"
daterenamer = "essembeh_tools.cli.daterenamer:run"
batxaran = "essembeh_tools.cli.batxaran:run"
remote-borg = "essembeh_tools.cli.remoteborg:run"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
